[ecash-herald] Patch use of float vs int for sats
Summary: Oversight from implementation of staking rewards parsing. `parseFloat` will work fine but we should `parseInt` as all `output.value` instances from chronik are expected to be strings of satoshi values, i.e. integers.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14791

+++ apps/ecash-herald/src/parse.js
@@ -106,7 +106,7 @@ module.exports = {
         const STAKING_REWARDS_PERCENT = 10;
         const { outputs } = coinbaseTx;
         const totalCoinbaseSats = outputs
-            .map(output => parseFloat(output.value))
+            .map(output => parseInt(output.value))
             .reduce((prev, curr) => prev + curr, 0);
         for (let output of outputs) {
             const thisValue = parseInt(output.value);
